Add initial implementation for using libraries
Basically, now doing `import foo::bar` expects `foo` to be a *library*.
We will search for a directory called `foo` in the list of library
paths, and import it. By default, the library path is OCEN_ROOT,
(or cwd if not provided). You can add additional library paths with
the `-l path` through the command line