Classify reference types
Before we treated all references the same for LSP refs and renames.
This is not correct - for `Find all References` mode, we want to show
_every_ reference that _uses_ a symbol. For `Rename Symbol` mode, we
only want the usages that _explicitly_ name the symbol, and need to
be renamed.

We now track the types of each reference in the compiler and only
output information for the ones we care about in LSP mode