Finally - fix some annoying import issues :)
We were really not properly handling imports for cases with a directory
having a `mod.oc` file. This is partially due to the awkward way in
which we type-checked imports, we would explicitly try to handle the
first part of the import, and then loop over the rest.

However, when trying to import something from a `mod.oc` file in the
current directory, this would cause all of these to break for some
reason(s):

import .{ Foo, Bar } .     // Should work :(
import .mod::{ Foo, Bar }  // expected, can't import `mod`
import @a::b::f::{Foo,Bar} // Specify path from root, annoying

Now, the first option works, and is the nicest one IMO.

As a free-benefit, this re-factor has also improved LSP completions,
and now they get triggered as soon as you type `.` or `@`, which is
cool!