Initial implementation of Value Enums
To not break existing enums, these are temporarily declared with
`venum` keyword. They have a bunch of limitations currently, but some
basic examples work.

In the future, ideally all enums withh be replaced with Value Enums,
since they can also behave like regular ones if no fields are specified
for any of the variants. The only issue is that we should probably
codegen a native C enum in this case for easier comparisons etc.