Change AST representation of match cases
Previously the match statement node contained a list of MatchCond
objects. Each of these would have a condition, but an optional body.
All consecutive MatchCond until you saw one with a body were supposed
to be part of the same "branch". For eg:
    match foo {
      Foo | Bar | Baz => print(...)
      Qux => 1
    }
would be
    (Foo, null), (Bar, null), (Baz, print(...)), (Qux, 1)
This was a little awkward to traverse. Now, each MatchCond stores
a list of all the actual conditions, and also the body, which is a more
natural way of storing the AST.