Change LSP testing to check for subset in output.
Instead of always checking if the actual output is _exactly_ the
same as the expected output, this change allows you to now specify
certain _types_ of things you are looking for. For instance, if
you expect a dictionary with a specific key-value foo:bar, but don't
care about any of the other elements in the dictionary, you can do:

expected: {"foo":"bar"}

which will ignore the existence of any other keys.