Add new concept of "compiler operations" + @new
"Compiler Operations" are prefixed by @, they are not functions and
just compile-time syntax sugar for common operations. In the instance
of @new, it is a shorthand for taking the expression, creating enough
space in the heap to put in the result, and storing the result there,
and returning that pointer.